wk04

+++ function-2-1.cpp
@@ -1,16 +1,16 @@
 #include<iostream>
 
 float add_op(float left, float right){
-	int a=left+right;
+	float a=left+right;
 	return a;
 }
 
 float subtract_op(float left, float right){
-	int a=left-right;
+	float a=left-right;
 	return a;
 }
 float arithmetic_ops(float left, float right, std::string op){
-	int a=0;
+	float a=0;
 	if(op=="add"){
 		 a=add_op(left,right);
 	}else if(op=="subtract"){
